Fix blueprint endpoint names in templates
