+++ Dockerfile
@@ -7,7 +7,8 @@ WORKDIR /app
 COPY requirements.txt .
 RUN pip install --no-cache-dir -r requirements.txt
 
-COPY app.py .
+COPY app.py routes.py ./
+COPY services ./services
 COPY templates ./templates
 COPY config ./config
