+++ templates/index.html
@@ -28,7 +28,7 @@
       </tbody>
     </table>
     {% else %}
-    <div class="alert alert-info">No hay datos de configuración. Ingresá a <a href="{{ url_for('admin') }}">/adm</a> para cargarlos.</div>
+    <div class="alert alert-info">No hay datos de configuración. Ingresá a <a href="{{ url_for('app.admin') }}">/adm</a> para cargarlos.</div>
     {% endif %}
   </div>
 </body>
